kvfollowerreadsccl: move to pkg/kv/followerreads
Epic: none
Release note: None
