+++ pkg/BUILD.bazel
@@ -49,7 +49,6 @@ ALL_TESTS = [
     "//pkg/ccl/cloudccl/gcp:gcp_test",
     "//pkg/ccl/importerccl:importerccl_test",
     "//pkg/ccl/jwtauthccl:jwtauthccl_test",
-    "//pkg/ccl/kvccl/kvfollowerreadsccl:kvfollowerreadsccl_test",
     "//pkg/ccl/ldapccl:ldapccl_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:3node-tenant-multiregion_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant:3node-tenant_test",
@@ -240,6 +239,7 @@ ALL_TESTS = [
     "//pkg/keyvisualizer/spanstatsconsumer:spanstatsconsumer_test",
     "//pkg/kv/bulk/bulkpb:bulkpb_test",
     "//pkg/kv/bulk:bulk_test",
+    "//pkg/kv/followerreads:followerreads_test",
     "//pkg/kv/kvclient/kvcoord/txncorrectnesstest:txncorrectnesstest_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord_disallowed_imports_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord_test",
@@ -1019,9 +1019,6 @@ GO_TARGETS = [
     "//pkg/ccl/importerccl:importerccl_test",
     "//pkg/ccl/jwtauthccl:jwtauthccl",
     "//pkg/ccl/jwtauthccl:jwtauthccl_test",
-    "//pkg/ccl/kvccl/kvfollowerreadsccl:kvfollowerreadsccl",
-    "//pkg/ccl/kvccl/kvfollowerreadsccl:kvfollowerreadsccl_test",
-    "//pkg/ccl/kvccl:kvccl",
     "//pkg/ccl/ldapccl:ldapccl",
     "//pkg/ccl/ldapccl:ldapccl_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:3node-tenant-multiregion_test",
@@ -1603,6 +1600,8 @@ GO_TARGETS = [
     "//pkg/kv/bulk/bulkpb:bulkpb_test",
     "//pkg/kv/bulk:bulk",
     "//pkg/kv/bulk:bulk_test",
+    "//pkg/kv/followerreads:followerreads",
+    "//pkg/kv/followerreads:followerreads_test",
     "//pkg/kv/kvbase:kvbase",
     "//pkg/kv/kvclient/kvcoord/txncorrectnesstest:txncorrectnesstest_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord",
