+++ pkg/backup/BUILD.bazel
@@ -51,7 +51,6 @@ go_library(
         "//pkg/backup/backuputils",
         "//pkg/base",
         "//pkg/build",
-        "//pkg/ccl/kvccl/kvfollowerreadsccl",
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/storageccl",
         "//pkg/cloud",
@@ -67,6 +66,7 @@ go_library(
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/kv/bulk",
+        "//pkg/kv/followerreads",
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver/batcheval",
         "//pkg/kv/kvserver/concurrency/lock",
@@ -231,7 +231,6 @@ go_test(
         "//pkg/blobs",
         "//pkg/build",
         "//pkg/build/bazel",
-        "//pkg/ccl/kvccl",
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/multiregionccl/multiregionccltestutils",
         "//pkg/ccl/multitenantccl",
@@ -253,6 +252,7 @@ go_test(
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/kv/bulk",
+        "//pkg/kv/followerreads",
         "//pkg/kv/kvclient/kvcoord",
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver",
