+++ pkg/ccl/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
         "//pkg/ccl/cliccl",
         "//pkg/ccl/gssapiccl",
         "//pkg/ccl/jwtauthccl",
-        "//pkg/ccl/kvccl",
         "//pkg/ccl/ldapccl",
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/multitenantccl",
@@ -28,6 +27,7 @@ go_library(
         "//pkg/crosscluster/logical",
         "//pkg/crosscluster/physical",
         "//pkg/crosscluster/producer",
+        "//pkg/kv/followerreads",
         "//pkg/server/license",
     ],
 )
