+++ pkg/ccl/changefeedccl/BUILD.bazel
@@ -66,7 +66,6 @@ go_library(
         "//pkg/ccl/changefeedccl/schemafeed",
         "//pkg/ccl/changefeedccl/tableset",
         "//pkg/ccl/changefeedccl/timers",
-        "//pkg/ccl/kvccl/kvfollowerreadsccl",
         "//pkg/ccl/utilccl",
         "//pkg/cloud",
         "//pkg/cloud/externalconn",
@@ -82,6 +81,7 @@ go_library(
         "//pkg/jobs/jobsprotectedts",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/kv/followerreads",
         "//pkg/kv/kvclient/kvcoord",
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver",
