+++ pkg/ccl/importerccl/BUILD.bazel
@@ -18,13 +18,13 @@ go_test(
         "//pkg/base",
         "//pkg/build/bazel",
         "//pkg/ccl",
-        "//pkg/ccl/kvccl",
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/multiregionccl/multiregionccltestutils",
         "//pkg/ccl/partitionccl",
         "//pkg/cloud/impl:cloudimpl",
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
+        "//pkg/kv/followerreads",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
