--- pkg/ccl/kvccl/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "kvccl",
-    srcs = ["kvccl.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/kvccl",
-    visibility = ["//visibility:public"],
-    deps = ["//pkg/ccl/kvccl/kvfollowerreadsccl"],
-)
