+++ pkg/ccl/serverccl/BUILD.bazel
@@ -47,10 +47,10 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/ccl/kvccl",
         "//pkg/ccl/utilccl/licenseccl",
         "//pkg/clusterversion",
         "//pkg/jobs",
+        "//pkg/kv/followerreads",
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver/liveness",
         "//pkg/kv/kvserver/liveness/livenesspb",
