+++ pkg/crosscluster/producer/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/crosscluster/producer",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/ccl/kvccl/kvfollowerreadsccl",
         "//pkg/ccl/utilccl",
         "//pkg/clusterversion",
         "//pkg/crosscluster",
@@ -27,6 +26,7 @@ go_library(
         "//pkg/jobs/jobsprotectedts",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/kv/followerreads",
         "//pkg/kv/kvclient/rangefeed",
         "//pkg/kv/kvclient/rangefeed/rangefeedcache",
         "//pkg/kv/kvpb",
