+++ pkg/kv/followerreads/BUILD.bazel
@@ -1,15 +1,14 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "kvfollowerreadsccl",
+    name = "followerreads",
     srcs = [
         "boundedstaleness.go",
         "followerreads.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvfollowerreadsccl",
+    importpath = "github.com/cockroachdb/cockroach/pkg/kv/followerreads",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/ccl/utilccl",
         "//pkg/kv",
         "//pkg/kv/kvclient/kvcoord",
         "//pkg/kv/kvpb",
@@ -34,7 +33,7 @@ go_library(
 )
 
 go_test(
-    name = "kvfollowerreadsccl_test",
+    name = "followerreads_test",
     size = "medium",
     srcs = [
         "boundedstaleness_test.go",
@@ -42,14 +41,12 @@ go_test(
         "main_test.go",
     ],
     data = glob(["testdata/**"]),
-    embed = [":kvfollowerreadsccl"],
+    embed = [":followerreads"],
     exec_properties = select({
         "//conditions:default": {"test.Pool": "large"},
     }),
     deps = [
         "//pkg/base",
-        "//pkg/ccl",
-        "//pkg/ccl/utilccl",
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/kv/kvbase",
